requirements.txt
streamlit
pandas
numpy
plotly
altair
matplotlib
python-dotenv
openai
textblob